+++ runtime/dummychecking.js
@@ -1,3 +1,0 @@
-exports.checkConfig = function(){
-  // TODO: Finish this function :)
-};